Extract schedule grid to eventer/schedule.py; coordinator schedule view now uses hourly grid with role colors and display names
